move type definitions to types package
